tax_clustering/requirements.txt
Django>=4.0,<5.0
gunicorn==20.1.0
djangorestframework
psycopg2-binary
pandas
numpy
scikit-learn
umap-learn
seaborn
matplotlib
statsmodels
scipy
django-cors-headers
gunicorn
tensorflow
umap